requirements.txt
streamlit
openai
google-generativeai
python-dotenv
PyPDF2
python-docx
Pillow